% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EARS-package.R
\docType{package}
\name{EARS}
\alias{EARS-package}
\alias{EARS}
\title{EARS: Event-Adjusted Rank Sum Test}
\description{
A non-parametric survival test that adjusts event times by the proportion of
events in each group and applies a Kruskal–Wallis test with a censoring penalty.
}
\author{
\strong{Maintainer}: Gustav Stålhammar \email{gustav.stalhammar@ki.se}

}
